Ignore the channel argument to the scene page touch function
In the current implementation of multichannel image rendering, all left eye pages are requested and drawn before all right eye pages. But because the request queue has a limited capacity (it's a traditional semaphored P-C queue) the left eye rendering starves the right eye. The user ends up seeing full resolution in the left eye while still seeing black in the right eye.

To alleviate this, any request in any channel triggers a request for the same page in all channels. This works well in all current cases. It is not, however, an elegant solution for the future.

The best solution is *not* to increase the capacity of the request queue. This would only incur wasted effort when the user changes scenes before the queue has emptied. Low-priority pages in the first scene would still be loaded *after* high-priority pages in the second scene, only to be discarded.

A better approach would be to merge the total request for all channels in the prep phase before proceeding to the draw phase. For this, the `pages` mechanism in the SCM sphere must be generalized.